refactor: renames zone list views
Renames the existing zone routes for list views to follow the established `*-list-view` convention.
