+++ src/app/zones/views/ZoneEgressListView.vue
