+++ src/app/zones/views/ZoneListView.vue
